requirements.txt
# Drone Final Year Project - Laptop Training Dependencies
# Install: pip install -r requirements.txt

# Core ML
ultralytics>=8.1.0          # YOLOv8 framework
torch>=2.0.0                # PyTorch (CUDA support for RTX 3050)
torchvision>=0.15.0

# Dataset
kagglehub>=0.2.0            # Download datasets from Kaggle

# Data & Visualization
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
pandas>=2.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
